consumoor: fix lint issues from CI
- Preallocate pooledEvents slice, fix wsl whitespace
- Suppress nilnil on intentional nil reject sink returns
- Name return values on LogSampler.Allow (gocritic)
- Remove unnecessary uint64 conversion after schema migration
  changed DurationMs columns from UInt32 to UInt64
- Add blank line before range loop (wsl)

+++ pkg/consumoor/route/execution/execution_engine_get_blobs.go
@@ -90,7 +90,7 @@ func (b *executionEngineGetBlobsBatch) appendPayload(event *xatu.DecoratedEvent)
 	}
 
 	if durationMs := payload.GetDurationMs(); durationMs != nil {
-		b.DurationMs.Append(uint64(durationMs.GetValue()))
+		b.DurationMs.Append(durationMs.GetValue())
 	} else {
 		b.DurationMs.Append(0)
 	}
