+++ pkg/consumoor/route/execution/execution_engine_new_payload.go
@@ -102,7 +102,7 @@ func (b *executionEngineNewPayloadBatch) appendPayload(event *xatu.DecoratedEven
 	}
 
 	if durationMs := payload.GetDurationMs(); durationMs != nil {
-		b.DurationMs.Append(uint64(durationMs.GetValue()))
+		b.DurationMs.Append(durationMs.GetValue())
 	} else {
 		b.DurationMs.Append(0)
 	}
